Move schedule to models
