+++ src/virtualship/expedition/__init__.py
@@ -2,7 +2,7 @@
 
 from .do_expedition import do_expedition
 from .input_data import InputData
-from .schedule import Schedule, Waypoint
+from ..models.schedule import Schedule, Waypoint
 from ..models.ship_config import (
     ADCPConfig,
     ArgoFloatConfig,
