+++ src/virtualship/models/schedule.py
@@ -11,12 +11,12 @@ import pydantic
 import pyproj
 import yaml
 
-from ..models.location import Location
-from ..models.ship_config import InstrumentType
-from .space_time_region import SpaceTimeRegion
+from .location import Location
+from .ship_config import InstrumentType
+from ..expedition.space_time_region import SpaceTimeRegion
 
 if TYPE_CHECKING:
-    from .input_data import InputData
+    from ..expedition.input_data import InputData
     from parcels import FieldSet
 
 projection: pyproj.Geod = pyproj.Geod(ellps="WGS84")
