+++ tests/expedition/test_ship_config.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from virtualship.expedition.schedule import Schedule
+from virtualship.models.schedule import Schedule
 from virtualship.models.ship_config import ConfigError, ShipConfig
 from virtualship.utils import get_example_config, get_example_schedule
