+++ tests/test_mfp_to_yaml.py
@@ -3,7 +3,7 @@ import os
 import pandas as pd
 import pytest
 
-from virtualship.expedition.schedule import Schedule
+from virtualship.models.schedule import Schedule
 from virtualship.models.ship_config import InstrumentType
 from virtualship.utils import mfp_to_yaml
